Dec 7-8/08-12-2024-7adabba9406c15d80ddbf1107f8cd459/Plot_CNN_op/requirements.txt
torch==2.0.1+cu118
torchvision==0.15.2+cu118
Pillow
matplotlib

# pip install -r requirements.txt --index-url https://download.pytorch.org/whl/cu118

# conda activate pytorch_gpu11_8